requirements.txt
ultralytics==8.0.20
streamlit==1.20.0
pillow==9.5.0
requests==2.31.0
opencv-python-headless==4.7.0.72
torch==2.0.1